refactor: CustomGenerator 의 위치 수정
