+++ src/test/java/domain/ResultTest.java
@@ -3,7 +3,6 @@ package domain;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import util.CustomGenerator;
 
 import java.util.List;
 import java.util.Map;
